pyproject.toml
[project]
name = "regmed-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp>=3.13.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "supabase>=2.27.1",
    "unstructured[docx,pdf]>=0.18.26",
]
